// src/types/MySQLType.ts
export type MySQLType =
  | 'INT'
  | 'SMALLINT'
  | 'TINYINT'
  | 'MEDIUMINT'
  | 'BIGINT'
  | 'FLOAT'
  | 'DOUBLE'
  | 'DECIMAL'
  | 'CHAR'
  | 'VARCHAR'
  | 'TEXT'
  | 'TINYTEXT'
  | 'MEDIUMTEXT'
  | 'LONGTEXT'
  | 'BOOLEAN'
  | 'TINYINT(1)'
  | 'DATE'
  | 'DATETIME'
  | 'TIMESTAMP'
  | 'TIME'
  | 'YEAR'
  | 'BINARY'
  | 'VARBINARY'
  | 'BLOB'
  | 'TINYBLOB'
  | 'MEDIUMBLOB'
  | 'LONGBLOB'
  | 'ENUM'
  | 'SET'
  | 'JSON'
  | 'GEOMETRY'
  | 'POINT'
  | 'LINESTRING'
  | 'POLYGON';